+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Security   to invite users to upgrade in case of vulnerabilities.
 -->
 
 ## [Unreleased]
+
+## [0.3.6]
 ### Added
 - New method `Multiline::getLineIndexByCharacterPosition` returns the line index containing a certain position.
 
@@ -117,7 +119,8 @@ Security   to invite users to upgrade in case of vulnerabilities.
 ### Added
 - First public release.
 
-[Unreleased]: https://github.com/nochso/omni/compare/0.3.5...HEAD
+[Unreleased]: https://github.com/nochso/omni/compare/0.3.6...HEAD
+[0.3.6]: https://github.com/nochso/omni/compare/0.3.5...0.3.6
 [0.3.5]: https://github.com/nochso/omni/compare/0.3.4...0.3.5
 [0.3.4]: https://github.com/nochso/omni/compare/0.3.3...0.3.4
 [0.3.3]: https://github.com/nochso/omni/compare/0.3.2...0.3.3